requirements.txt
google-cloud-bigquery>=3.13.0
google-auth>=2.22.0
pandas>=2.0.0
db-dtypes>=1.0.0
jupyter>=1.0.0